Rename package Resource to ResourceModel for PHP7 support
