+++ Model/ResourceModel/Popup.php
@@ -9,7 +9,7 @@
  * @license     http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 
-namespace Ebizmarts\AbandonedCart\Model\Resource;
+namespace Ebizmarts\AbandonedCart\Model\ResourceModel;
 
 class Popup extends \Magento\Framework\Model\Resource\Db\AbstractDb
 {
